refactor: name to the showToast and merge showerrortoast and showsuccesstoast
