+++ src/app/(private)/roles/create-role/page.tsx
@@ -13,7 +13,7 @@ import {
 import { Input } from '@/components/ui/input'
 import { Switch } from '@/components/ui/switch'
 import { useToast } from '@/hooks/useToast'
-import { showErrorToast } from '@/lib/showErrorToast'
+import { showErrorToast } from '@/lib/showToast'
 import {
   getLocalizedCategory,
   getLocalizedPermission,
