+++ src/components/dashboard/menu.tsx
@@ -12,7 +12,7 @@ import { LoadingSpinner } from '@/components/ui/loadingSpinner'
 import { selectPermissions } from '@/modules/auth/authSlice'
 import { useAppSelector } from '@/store/hooks'
 import { Permission } from '@/constants/permissions'
-import { showErrorToast } from '@/lib/showErrorToast'
+import { showErrorToast } from '@/lib/showToast'
 
 const Menu = (): JSX.Element => {
   const pathname = usePathname()
